requirements.txt
click==8.0.3
Flask==0.12.5
Flask-APIBlueprint==1.0.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
PyJWT==2.3.0
python-dotenv==0.19.2
six==1.16.0
typing_extensions==4.0.1
Werkzeug==0.16.1
zipp==3.6.0
